requirements.txt
pycryptodome==3.15.0
pillow==9.2.0
qrcode==7.3.1